requirements.txt
minio==7.1.12
munch==2.5.0
natsort==8.2.0
openstacksdk==0.101.0
paramiko==2.11.0
patool==1.12
requests==2.28.1
tabulate==0.8.10
typer==0.6.1
